rename: api 하위로 도메인 디렉토리 변경
