+++ be/src/main/java/kr/codesquad/secondhand/api/jwt/domain/MemberRefreshToken.java
@@ -1,0 +1,23 @@
+package kr.codesquad.secondhand.api.jwt.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Table(name = "token")
+public class MemberRefreshToken {
+
+    @Id
+    private Long memberId;
+
+    private String refreshToken;
+
+}
