+++ be/src/main/java/kr/codesquad/secondhand/api/jwt/repository/TokenRepository.java
@@ -1,0 +1,13 @@
+package kr.codesquad.secondhand.api.jwt.repository;
+
+import kr.codesquad.secondhand.api.jwt.domain.MemberRefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TokenRepository extends JpaRepository<MemberRefreshToken, Long> {
+
+//    boolean existsById(Long memberId);
+
+//    void deleteById(Long memberId);
+}
