+++ be/src/main/java/kr/codesquad/secondhand/api/member/repository/JpaMemberRepository.java
@@ -1,0 +1,10 @@
+package kr.codesquad.secondhand.api.member.repository;
+
+import java.util.Optional;
+import kr.codesquad.secondhand.api.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaMemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findBySignInTypeIdAndEmail(Long signInTypeId, String email);
+}
