+++ be/src/main/java/kr/codesquad/secondhand/api/member/repository/MemberAddressRepositoryImpl.java
@@ -1,0 +1,11 @@
+package kr.codesquad.secondhand.api.member.repository;
+
+import java.util.List;
+import java.util.Optional;
+import kr.codesquad.secondhand.api.member.domain.MemberAddress;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberAddressRepositoryImpl extends JpaRepository<MemberAddress, Long> {
+
+    Optional<List<MemberAddress>> findAllByMemberId(Long memberId);
+}
