+++ be/src/main/java/kr/codesquad/secondhand/api/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package kr.codesquad.secondhand.api.member.repository;
+
+import java.util.Optional;
+import kr.codesquad.secondhand.api.member.domain.Member;
+
+public interface MemberRepository {
+
+    Long save(Member member);
+
+    Optional<Member> findBySignInTypeIdAndEmail(Long signInTypeId, String email);
+
+}
