+++ be/src/main/java/kr/codesquad/secondhand/api/oauth/dto/OAuthTokenResponse.java
@@ -1,0 +1,26 @@
+package kr.codesquad.secondhand.api.oauth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class OAuthTokenResponse {
+
+    @JsonProperty("access_token")
+    private String accessToken;
+
+    private String scope;
+
+    @JsonProperty("token_type")
+    private String tokenType;
+
+    @Builder
+    public OAuthTokenResponse(String accessToken, String scope, String tokenType) {
+        this.accessToken = accessToken;
+        this.scope = scope;
+        this.tokenType = tokenType;
+    }
+}
