--- be/src/main/java/kr/codesquad/secondhand/jwt/domain/Jwt.java
@@ -1,13 +1,0 @@
-package kr.codesquad.secondhand.jwt.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Jwt {
-
-    private String accessToken;
-    private String refreshToken;
-
-}
