--- be/src/main/java/kr/codesquad/secondhand/jwt/domain/JwtProperties.java
@@ -1,18 +1,0 @@
-package kr.codesquad.secondhand.jwt.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@Component
-@ConfigurationProperties("jwt")
-public class JwtProperties {
-
-    private String secretKey;
-    private Long accessTokenExpirationTime;
-    private Long refreshTokenExpirationTime;
-
-}
