--- be/src/main/java/kr/codesquad/secondhand/jwt/domain/MemberRefreshToken.java
@@ -1,23 +1,0 @@
-package kr.codesquad.secondhand.jwt.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Table(name = "token")
-public class MemberRefreshToken {
-
-    @Id
-    private Long memberId;
-
-    private String refreshToken;
-
-}
