--- be/src/main/java/kr/codesquad/secondhand/member/domain/Member.java
@@ -1,47 +1,0 @@
-package kr.codesquad.secondhand.member.domain;
-
-import java.time.LocalDateTime;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EntityListeners(AuditingEntityListener.class)
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String email;
-    private String nickname;
-    private String profileImgUrl;
-
-    @CreatedDate
-    private LocalDateTime createdTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "sign_in_type_id")
-    private SignInType signInType;
-
-    @Builder
-    public Member(SignInType signInType, String email, String nickname, String profileImgUrl) {
-        this.signInType = signInType;
-        this.email = email;
-        this.nickname = nickname;
-        this.profileImgUrl = profileImgUrl;
-    }
-}
