--- be/src/main/java/kr/codesquad/secondhand/member/domain/MemberAddress.java
@@ -1,38 +1,0 @@
-package kr.codesquad.secondhand.member.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberAddress {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "address_id")
-    private Address address;
-
-    private boolean isLastVisited;
-
-    public MemberAddress(Member member, Address address, boolean isLastVisited) {
-        this.member = member;
-        this.address = address;
-        this.isLastVisited = isLastVisited;
-    }
-}
