--- be/src/main/java/kr/codesquad/secondhand/member/repository/AddressRepositoryImpl.java
@@ -1,9 +1,0 @@
-package kr.codesquad.secondhand.member.repository;
-
-import kr.codesquad.secondhand.member.domain.Address;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AddressRepositoryImpl extends JpaRepository<Address, Long> {
-}
