--- be/src/main/java/kr/codesquad/secondhand/member/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package kr.codesquad.secondhand.member.repository;
-
-import java.util.Optional;
-import kr.codesquad.secondhand.member.domain.Member;
-
-public interface MemberRepository {
-
-    Long save(Member member);
-
-    Optional<Member> findBySignInTypeIdAndEmail(Long signInTypeId, String email);
-
-}
